添加清除缓存

+++ app/src/main/java/com/monke/monkeybook/help/BookshelfHelp.java
@@ -93,6 +93,12 @@ public class BookshelfHelp {
         return chapterCaches.containsKey(path) && chapterCaches.get(path).contains(chapter.getDurChapterIndex());
     }
 
+    public static void clearCaches() {
+        FileHelp.deleteFile(Constant.BOOK_CACHE_PATH);
+        FileHelp.getFolder(Constant.BOOK_CACHE_PATH);
+        chapterCaches.clear();
+    }
+
     /**
      * 删除章节文件
      */
@@ -183,12 +189,18 @@ public class BookshelfHelp {
         return null;
     }
 
-    public static void removeFromBookShelf(BookShelfBean bookShelfBean) {
+    public static void removeFromBookShelf(BookShelfBean bookShelfBean, boolean keepCaches) {
         DbHelper.getInstance().getmDaoSession().getBookShelfBeanDao().deleteByKey(bookShelfBean.getNoteUrl());
         DbHelper.getInstance().getmDaoSession().getBookInfoBeanDao().deleteByKey(bookShelfBean.getBookInfoBean().getNoteUrl());
         DbHelper.getInstance().getmDaoSession().getChapterListBeanDao().deleteInTx(bookShelfBean.getChapterList());
-        FileHelp.deleteFile(Constant.BOOK_CACHE_PATH + getCachePathName(bookShelfBean.getBookInfoBean()));
-        chapterCaches.remove(getCachePathName(bookShelfBean.getBookInfoBean()));
+        if(!keepCaches) {
+            FileHelp.deleteFile(Constant.BOOK_CACHE_PATH + getCachePathName(bookShelfBean.getBookInfoBean()));
+            chapterCaches.remove(getCachePathName(bookShelfBean.getBookInfoBean()));
+        }
+    }
+
+    public static void removeFromBookShelf(BookShelfBean bookShelfBean) {
+        removeFromBookShelf(bookShelfBean, false);
     }
 
     public static void saveBookToShelf(BookShelfBean bookShelfBean) {
